Balance type changed to BigDecimal

+++ src/main/java/com/cristianregnier/accountingnotebook/api/AccountController.java
@@ -5,6 +5,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
+
 @RestController
 public class AccountController {
 
@@ -12,7 +14,7 @@ public class AccountController {
    private AccountManager accountManager;
 
    @GetMapping("/")
-   public Double getAccountBalance() {
+   public BigDecimal getAccountBalance() {
       return accountManager.getAccountBalance();
    }
 }
