+++ src/main/java/com/cristianregnier/accountingnotebook/api/command/CommitTransactionCommand.java
@@ -3,12 +3,14 @@ package com.cristianregnier.accountingnotebook.api.command;
 import com.cristianregnier.accountingnotebook.model.Transaction;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class CommitTransactionCommand {
    private Transaction.Type type;
-   private Double amount;
+   private BigDecimal amount;
 
-   public CommitTransactionCommand(Transaction.Type type, Double amount) {
+   public CommitTransactionCommand(Transaction.Type type, BigDecimal amount) {
       if (type == null)
          throw new IllegalArgumentException("Operation type is required.");
       if (amount == null)
